Split home page into an internal and an external view of the homepage
