+++ src/pages/Home/HomePageInternal.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import bannerImage from '../../assets/landingBannerImage.svg';
+import telegramLogo from '../../assets/telegramLogo.svg';
+import twitterLogo from '../../assets/twitterLogo.svg';
+import discordLogo from '../../assets/discordLogo.svg';
+import defaultUserProfileImage from '../../assets/defaultUserProfileImage.png';
+
+function HomePageInternal(): JSX.Element {
+  return <div>Internal Home Page</div>;
+}
+
+export default HomePageInternal;
